Missing Images. Still WIP on game logic
Added the images that were missed in a merge. Game logic is mostly
working but has a few bugs, mostly with users disconnecting during the
game. Also, there's a bug with having the top card revealed and an empty
pile. If the previous player passes a card, it isn't added to the pile.
I think I know what the problem is though.