Game is playable. Not all features done yet.
You can play a game with friends now and it should hopefully not cause
any errors, even if disconnects occur. There are still a handful of
things to add such as penalty timers and player stats as well as
indicating that a player disconnected and making the "YOU WIN/LOSE" font
bigger (placeholder for now to see if everything was working).