Last minute things. Should be final.
Stats work and are displayed in header. John's fix to our server crash
had flaws in it and were corrected. Some styling changes. Also made it
where a spoon graphic is drawn in a user's div box if they got a spoon.
Count down timer is more obvious. And not sure if it works, but I added
a pre-load images thing into the login and register pages.